# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse API
  version: 0.1.0
  description: |
    Tenant-scoped event ingestion and analytics API. The OpenAPI document
    describes the routes currently registered by the Go HTTP handler. It is
    not a claim that all routes have passed an end-to-end execution.
servers:
  - url: http://localhost:8080
    description: Ingestion service in Docker Compose
  - url: http://localhost:8081
    description: Query API in Docker Compose
tags:
  - name: Events
  - name: Tenants
  - name: Webhooks
  - name: Analytics
  - name: Health
paths:
  /health/live:
    get:
      tags: [Health]
      summary: Liveness probe
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Health' }
  /health/ready:
    get:
      tags: [Health]
      summary: Readiness probe
      responses:
        '200':
          description: Process is ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Health' }
  /metrics:
    get:
      tags: [Health]
      summary: Prometheus metrics
      responses:
        '200':
          description: Prometheus exposition format
          content:
            text/plain: {}
  /v1/events:
    post:
      tags: [Events]
      summary: Publish an event
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
            example:
              eventId: evt_8fa91
              type: payment.completed
              timestamp: '2026-09-09T22:30:00Z'
              payload: { paymentId: '123', amount: 499.9 }
      responses:
        '202':
          description: Published to the configured Kafka producer durability point
          headers:
            X-Pulse-Durability:
              schema: { type: string, example: kafka-ack }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedEvent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/events/batch:
    post:
      tags: [Events]
      summary: Publish multiple events at one Kafka durability point
      description: The whole batch is acknowledged only after all events are durably published. The default request body limit is 16 MiB and the event limit is 500; deployments may configure both limits.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 500
              items: { $ref: '#/components/schemas/EventInput' }
      responses:
        '202':
          description: All events were published to the configured Kafka producer durability point
          headers:
            X-Pulse-Durability:
              schema: { type: string, example: kafka-ack }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedBatch' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/tenants:
    post:
      tags: [Tenants]
      summary: Create a tenant
      security: [{ adminToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
      responses:
        '201':
          description: Tenant and one-time API key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/tenants/{tenantId}/rotate:
    post:
      tags: [Tenants]
      summary: Rotate the authenticated tenant API key
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: New one-time API key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KeyRotated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Create an event webhook subscription
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType, endpoint]
              properties:
                eventType: { type: string, minLength: 1 }
                endpoint: { type: string, format: uri, pattern: '^https?://' }
      responses:
        '201':
          description: Subscription and generated secret
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WebhookCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { $ref: '#/components/responses/NotImplemented' }
  /v1/analytics/summary:
    get:
      tags: [Analytics]
      summary: Count unique event IDs by type for a tenant and time range
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Inclusive RFC3339 start; defaults to 24 hours ago.
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Exclusive RFC3339 end; defaults to now.
        - in: query
          name: type
          schema: { type: string }
      responses:
        '200':
          description: Summary rows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AnalyticsSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '501': { $ref: '#/components/responses/NotImplemented' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  parameters:
    TenantId:
      in: path
      name: tenantId
      required: true
      schema: { type: string }
  schemas:
    EventInput:
      type: object
      additionalProperties: false
      required: [eventId, type, timestamp, payload]
      properties:
        eventId: { type: string, minLength: 1, maxLength: 200 }
        type: { type: string, minLength: 1, maxLength: 200, pattern: '^\\S+$' }
        timestamp: { type: string, format: date-time }
        payload:
          description: Any valid non-null JSON value within the configured payload limit.
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
            - type: string
            - type: number
            - type: boolean
    AcceptedEvent:
      type: object
      required: [eventId, status]
      properties:
        eventId: { type: string }
        status: { type: string, enum: [accepted] }
    AcceptedBatch:
      type: object
      required: [count, status]
      properties:
        count: { type: integer, minimum: 1 }
        status: { type: string, enum: [accepted] }
    TenantCreated:
      type: object
      required: [tenantId, name, apiKey]
      properties:
        tenantId: { type: string }
        name: { type: string }
        apiKey: { type: string, description: Returned only at creation. }
    KeyRotated:
      type: object
      required: [tenantId, apiKey]
      properties:
        tenantId: { type: string }
        apiKey: { type: string, description: Replaces the previous key. }
    WebhookCreated:
      type: object
      required: [id, eventType, endpoint, secret]
      properties:
        id: { type: string }
        eventType: { type: string }
        endpoint: { type: string, format: uri }
        secret: { type: string, description: Base64url secret returned at creation. }
    AnalyticsSummary:
      type: object
      required: [type, count]
      properties:
        type: { type: string }
        count: { type: integer, format: int64, minimum: 0 }
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Authentication failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated tenant does not own the resource
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Tenant rate limit exceeded
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unavailable:
      description: Required dependency or publish path unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotImplemented:
      description: Optional store is not configured
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }