# internal/api/http.go
package api

import (
	"context"
	"encoding/base64"
	"encoding/json"
	"io"
	"net"
	"net/http"
	"strings"
	"time"

	"github.com/prometheus/client_golang/prometheus"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"github.com/pulse-data/pulse/internal/analytics"
	"github.com/pulse-data/pulse/internal/auth"
	"github.com/pulse-data/pulse/internal/events"
	"github.com/pulse-data/pulse/internal/ratelimit"
	"github.com/pulse-data/pulse/internal/telemetry"
	"github.com/pulse-data/pulse/internal/webhook"
	"go.opentelemetry.io/otel/attribute"
	"go.opentelemetry.io/otel/codes"
	"go.opentelemetry.io/otel/propagation"
	"go.opentelemetry.io/otel/trace"
)

type Publisher interface {
	Publish(context.Context, events.Event) error
}
type BatchPublisher interface {
	PublishBatch(context.Context, []events.Event) error
}
type Handler struct {
	auth           auth.Store
	publisher      Publisher
	limiter        ratelimit.Limiter
	metrics        *telemetry.Metrics
	maxPayload     int
	maxBatchBytes  int
	maxBatchEvents int
	durability     string
	analytics      analytics.Store
	adminToken     string
	webhooks       webhook.Store
	metricsHTTP    http.Handler
	readyCheck     func(context.Context) error
}

func NewHandler(store auth.Store, publisher Publisher, limiter ratelimit.Limiter, metrics *telemetry.Metrics, maxPayload, maxBatchBytes, maxBatchEvents int, durability, adminToken string, analytical analytics.Store, webhooks webhook.Store, registry prometheus.Gatherer) *Handler {
	return &Handler{auth: store, publisher: publisher, limiter: limiter, metrics: metrics, maxPayload: maxPayload, maxBatchBytes: maxBatchBytes, maxBatchEvents: maxBatchEvents, durability: durability, analytics: analytical, adminToken: adminToken, webhooks: webhooks, metricsHTTP: promhttp.HandlerFor(registry, promhttp.HandlerOpts{})}
}

func (h *Handler) SetReadyCheck(check func(context.Context) error) { h.readyCheck = check }

func (h *Handler) Routes() http.Handler {
	mux := http.NewServeMux()
	mux.HandleFunc("GET /health/live", func(w http.ResponseWriter, _ *http.Request) {
		writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
	})
	mux.HandleFunc("GET /health/ready", func(w http.ResponseWriter, _ *http.Request) {
		if h.readyCheck != nil {
			ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
			defer cancel()
			if err := h.readyCheck(ctx); err != nil {
				h.fail(w, http.StatusServiceUnavailable, "dependencies are not ready")
				return
			}
		}
		writeJSON(w, http.StatusOK, map[string]string{"status": "ready"})
	})
	mux.Handle("GET /metrics", h.metricsHTTP)
	mux.HandleFunc("POST /v1/events", h.ingest)
	mux.HandleFunc("POST /v1/events/batch", h.ingestBatch)
	mux.HandleFunc("POST /v1/tenants", h.createTenant)
	mux.HandleFunc("POST /v1/tenants/", h.rotateTenant)
	mux.HandleFunc("POST /v1/webhooks", h.createWebhook)
	mux.HandleFunc("GET /v1/analytics/summary", h.summary)
	return h.instrument(mux)
}

func (h *Handler) ingest(w http.ResponseWriter, r *http.Request) {
	tenant, ok := h.authenticate(w, r)
	if !ok {
		return
	}
	if h.limiter != nil {
		allowed, err := h.limiter.Allow(r.Context(), tenant.ID)
		if err != nil {
			h.fail(w, http.StatusServiceUnavailable, "rate limiter unavailable")
			return
		}
		if !allowed {
			h.metrics.RateLimitRejected.WithLabelValues().Inc()
			h.fail(w, http.StatusTooManyRequests, "rate limit exceeded")
			return
		}
	}
	reader := http.MaxBytesReader(w, r.Body, int64(h.maxPayload)+4096)
	defer r.Body.Close()
	decoder := json.NewDecoder(reader)
	var raw json.RawMessage
	if err := decoder.Decode(&raw); err != nil {
		h.fail(w, http.StatusBadRequest, "invalid JSON")
		return
	}
	var extra any
	if err := decoder.Decode(&extra); err != io.EOF {
		h.fail(w, http.StatusBadRequest, "invalid JSON")
		return
	}
	event, err := events.Decode(raw, tenant.ID, h.maxPayload)
	if err != nil {
		h.metrics.EventsFailed.WithLabelValues("validation").Inc()
		h.fail(w, http.StatusBadRequest, err.Error())
		return
	}
	publishStarted := time.Now()
	if err := h.publisher.Publish(r.Context(), event); err != nil {
		h.metrics.EventsFailed.WithLabelValues("publish").Inc()
		h.fail(w, http.StatusServiceUnavailable, "event was not durable")
		return
	}
	h.metrics.KafkaPublishLatency.WithLabelValues("events.raw").Observe(time.Since(publishStarted).Seconds())
	h.metrics.Published.WithLabelValues("events.raw").Inc()
	h.metrics.EventsReceived.WithLabelValues().Inc()
	w.Header().Set("X-Pulse-Durability", h.durability)
	writeJSON(w, http.StatusAccepted, map[string]string{"eventId": event.EventID, "status": "accepted"})
}

func (h *Handler) ingestBatch(w http.ResponseWriter, r *http.Request) {
	tenant, ok := h.authenticate(w, r)
	if !ok {
		return
	}
	reader := http.MaxBytesReader(w, r.Body, int64(h.maxBatchBytes))
	defer r.Body.Close()
	var raw []json.RawMessage
	decoder := json.NewDecoder(reader)
	if err := decoder.Decode(&raw); err != nil || len(raw) == 0 || len(raw) > h.maxBatchEvents {
		h.fail(w, http.StatusBadRequest, "batch must contain between 1 and the configured maximum number of events")
		return
	}
	var extra any
	if err := decoder.Decode(&extra); err != io.EOF {
		h.fail(w, http.StatusBadRequest, "invalid JSON")
		return
	}
	batch := make([]events.Event, 0, len(raw))
	for _, value := range raw {
		event, err := events.Decode(value, tenant.ID, h.maxPayload)
		if err != nil {
			h.metrics.EventsFailed.WithLabelValues("validation").Inc()
			h.fail(w, http.StatusBadRequest, err.Error())
			return
		}
		batch = append(batch, event)
	}
	if allowed, err := h.allow(r.Context(), tenant.ID, len(batch)); err != nil {
		h.fail(w, http.StatusServiceUnavailable, "rate limiter unavailable")
		return
	} else if !allowed {
		h.metrics.RateLimitRejected.WithLabelValues().Inc()
		h.fail(w, http.StatusTooManyRequests, "rate limit exceeded")
		return
	}
	if publisher, ok := h.publisher.(BatchPublisher); ok {
		if err := publisher.PublishBatch(r.Context(), batch); err != nil {
			h.metrics.EventsFailed.WithLabelValues("publish").Add(float64(len(batch)))
			h.fail(w, http.StatusServiceUnavailable, "events were not durable")
			return
		}
	} else {
		for _, event := range batch {
			if err := h.publisher.Publish(r.Context(), event); err != nil {
				h.metrics.EventsFailed.WithLabelValues("publish").Inc()
				h.fail(w, http.StatusServiceUnavailable, "events were not durable")
				return
			}
		}
	}
	h.metrics.Published.WithLabelValues("events.raw").Add(float64(len(batch)))
	h.metrics.EventsReceived.WithLabelValues().Add(float64(len(batch)))
	w.Header().Set("X-Pulse-Durability", h.durability)
	writeJSON(w, http.StatusAccepted, map[string]any{"count": len(batch), "status": "accepted"})
}

func (h *Handler) allow(ctx context.Context, tenantID string, amount int) (bool, error) {
	if h.limiter == nil {
		return true, nil
	}
	if limiter, ok := h.limiter.(ratelimit.BatchLimiter); ok {
		return limiter.AllowN(ctx, tenantID, amount)
	}
	for i := 0; i < amount; i++ {
		allowed, err := h.limiter.Allow(ctx, tenantID)
		if err != nil || !allowed {
			return allowed, err
		}
	}
	return true, nil
}

func (h *Handler) authenticate(w http.ResponseWriter, r *http.Request) (auth.Tenant, bool) {
	key, ok := auth.Bearer(r.Header.Get("Authorization"))
	if !ok {
		h.fail(w, http.StatusUnauthorized, "missing bearer token")
		return auth.Tenant{}, false
	}
	tenant, err := h.auth.Authenticate(r.Context(), key)
	if err != nil {
		h.fail(w, http.StatusUnauthorized, "invalid credentials")
		return auth.Tenant{}, false
	}
	return tenant, true
}
func (h *Handler) createTenant(w http.ResponseWriter, r *http.Request) {
	if h.adminToken == "" || r.Header.Get("X-Admin-Token") != h.adminToken {
		h.fail(w, http.StatusUnauthorized, "admin authentication required")
		return
	}
	var in struct {
		Name string `json:"name"`
	}
	if json.NewDecoder(http.MaxBytesReader(w, r.Body, 4096)).Decode(&in) != nil {
		h.fail(w, http.StatusBadRequest, "invalid JSON")
		return
	}
	in.Name = strings.TrimSpace(in.Name)
	if in.Name == "" || len(in.Name) > 200 {
		h.fail(w, http.StatusBadRequest, "name must be between 1 and 200 characters")
		return
	}
	t, key, err := h.auth.Create(r.Context(), in.Name)
	if err != nil {
		h.fail(w, http.StatusBadRequest, err.Error())
		return
	}
	writeJSON(w, http.StatusCreated, map[string]any{"tenantId": t.ID, "name": t.Name, "apiKey": key})
}
func (h *Handler) rotateTenant(w http.ResponseWriter, r *http.Request) {
	parts := strings.Split(strings.Trim(r.URL.Path, "/"), "/")
	if len(parts) != 4 || parts[3] != "rotate" {
		h.fail(w, http.StatusNotFound, "not found")
		return
	}
	tenant, ok := h.authenticate(w, r)
	if !ok {
		return
	}
	if tenant.ID != parts[2] {
		h.fail(w, http.StatusForbidden, "tenant mismatch")
		return
	}
	key, err := h.auth.RotateKey(r.Context(), tenant.ID)
	if err != nil {
		h.fail(w, http.StatusNotFound, "tenant not found")
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"tenantId": tenant.ID, "apiKey": key})
}
func (h *Handler) summary(w http.ResponseWriter, r *http.Request) {
	if h.analytics == nil {
		h.fail(w, http.StatusNotImplemented, "analytics is not configured")
		return
	}
	tenant, ok := h.authenticate(w, r)
	if !ok {
		return
	}
	to := time.Now().UTC()
	from := to.Add(-24 * time.Hour)
	if value := r.URL.Query().Get("from"); value != "" {
		parsed, err := time.Parse(time.RFC3339, value)
		if err != nil {
			h.fail(w, http.StatusBadRequest, "invalid from")
			return
		} else {
			from = parsed
		}
	}
	if value := r.URL.Query().Get("to"); value != "" {
		parsed, err := time.Parse(time.RFC3339, value)
		if err != nil {
			h.fail(w, http.StatusBadRequest, "invalid to")
			return
		} else {
			to = parsed
		}
	}
	if !from.Before(to) {
		h.fail(w, http.StatusBadRequest, "from must be before to")
		return
	}
	if to.Sub(from) > 31*24*time.Hour {
		h.fail(w, http.StatusBadRequest, "analytics range cannot exceed 31 days")
		return
	}
	queryCtx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
	defer cancel()
	rows, err := h.analytics.Summary(queryCtx, tenant.ID, r.URL.Query().Get("type"), from, to)
	if err != nil {
		h.fail(w, http.StatusServiceUnavailable, "analytics unavailable")
		return
	}
	writeJSON(w, http.StatusOK, rows)
}
func (h *Handler) createWebhook(w http.ResponseWriter, r *http.Request) {
	if h.webhooks == nil {
		h.fail(w, http.StatusNotImplemented, "webhooks are not configured")
		return
	}
	tenant, ok := h.authenticate(w, r)
	if !ok {
		return
	}
	var in struct {
		EventType string `json:"eventType"`
		Endpoint  string `json:"endpoint"`
	}
	if json.NewDecoder(http.MaxBytesReader(w, r.Body, 4096)).Decode(&in) != nil || strings.TrimSpace(in.EventType) == "" {
		h.fail(w, http.StatusBadRequest, "eventType and endpoint are required")
		return
	}
	if err := webhook.ValidateEndpoint(r.Context(), in.Endpoint, func(ctx context.Context, host string) ([]net.IP, error) {
		return net.DefaultResolver.LookupIP(ctx, "ip", host)
	}); err != nil {
		h.fail(w, http.StatusBadRequest, "endpoint rejected by SSRF policy")
		return
	}
	id, err := webhook.NewID()
	if err != nil {
		h.fail(w, http.StatusInternalServerError, "could not create subscription")
		return
	}
	secret, err := webhook.NewSecret()
	if err != nil {
		h.fail(w, http.StatusInternalServerError, "could not create subscription")
		return
	}
	sub, err := h.webhooks.CreateSubscription(r.Context(), webhook.Subscription{ID: id, TenantID: tenant.ID, EventType: in.EventType, Endpoint: in.Endpoint, Secret: secret, Enabled: true, CreatedAt: time.Now().UTC(), UpdatedAt: time.Now().UTC()})
	if err != nil {
		h.fail(w, http.StatusInternalServerError, "could not create subscription")
		return
	}
	writeJSON(w, http.StatusCreated, map[string]any{"id": sub.ID, "eventType": sub.EventType, "endpoint": sub.Endpoint, "secret": base64.RawURLEncoding.EncodeToString(secret)})
}
func (h *Handler) instrument(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := propagation.TraceContext{}.Extract(r.Context(), propagation.HeaderCarrier(r.Header))
		ctx, span := telemetry.StartSpan(ctx, r.Method+" "+r.URL.Path, trace.WithSpanKind(trace.SpanKindServer), trace.WithAttributes(
			attribute.String("http.request.method", r.Method),
			attribute.String("url.path", r.URL.Path),
		))
		defer span.End()
		r = r.WithContext(ctx)
		started := time.Now()
		rw := &statusWriter{ResponseWriter: w, status: 200}
		next.ServeHTTP(rw, r)
		span.SetAttributes(attribute.Int("http.response.status_code", rw.status))
		if rw.status >= http.StatusInternalServerError {
			span.SetStatus(codes.Error, http.StatusText(rw.status))
		}
		route := r.Pattern
		if route == "" {
			route = "unknown"
		}
		h.metrics.Requests.WithLabelValues(r.Method, route, http.StatusText(rw.status)).Inc()
		h.metrics.RequestDuration.WithLabelValues(r.Method, route).Observe(time.Since(started).Seconds())
	})
}

type statusWriter struct {
	http.ResponseWriter
	status int
}

func (w *statusWriter) WriteHeader(status int) {
	w.status = status
	w.ResponseWriter.WriteHeader(status)
}
func (w *statusWriter) Write(b []byte) (int, error) { return w.ResponseWriter.Write(b) }
func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}
func (h *Handler) fail(w http.ResponseWriter, status int, message string) {
	writeJSON(w, status, map[string]string{"error": message})
}